More sementic analysis + updated Readme

+++ tests/CompilersBenchmark/SemanticErrors/declarationDefinitionMismatch2.c
@@ -2,7 +2,7 @@
 int f(int a, int b);
 
 // definition of f with wrong parameter type
-int f(int a, int b){
+int f(float a, int b){
     return b;
 }
